Server-side OLA API Booking integration and background jobs

+++ grails-app/domain/smart_transport/Location.groovy
@@ -7,7 +7,7 @@ class Location {
 
     static belongsTo = [user:User]
 
-    long latitude;
-    long longitude;
+    double latitude;
+    double longitude;
     String address; // Human readable address of lat/long
 }
